requirements.txt
Pillow
mysql-connector
datetime
ttkthemes
tkcalendar
matplotlib
scipy
